requirements.txt
# can pin versions later if needed
pandas
jupyter
matplotlib
pandas-profiling
plotly
flask
pyjade
pillow
